게시물 업로드 구조 개선

+++ src/post/post.service.ts
@@ -5,7 +5,6 @@ import { CreatePostDto } from './dto/create-post.dto';
 import { FindOneOptions, Repository } from 'typeorm';
 import { Post } from './entities/post.entity';
 import { User } from 'src/users/entities/user.entity';
-import { File as FileEntity } from 'src/files/entities/file.entity';
 import { FindPostsDto } from './dto/find-posts.dto';
 import { NullableType } from 'src/utils/types/nullable.type';
 import { UsersService } from 'src/users/users.service';
@@ -80,14 +79,12 @@ export class PostService {
     }
 
     async create(createPostDto: CreatePostDto, user: User): Promise<Post> {
-        const { title, content, description, status, categoryId, tagNames, imageIds } = createPostDto;
+        const { title, content, description, status, categoryId, tagNames, fileNames } = createPostDto;
 
         let category;
         if (categoryId) {
             category = await this.categoryService.findOneById(categoryId);
-            if (!category) {
-                throw new NotFoundException('존재하지 않는 카테고리입니다.');
-            }
+            if (!category) throw new NotFoundException('존재하지 않는 카테고리입니다.');
         }
 
         const tags: Tag[] = [];
@@ -98,15 +95,7 @@ export class PostService {
             }
         }
 
-        const images: FileEntity[] = [];
-        if (imageIds) {
-            for (const imageId of imageIds) {
-                const imageFile = await this.filesService.findById(imageId);
-                images.push(imageFile);
-            }
-        }
-
-        return this.postRepository.save(
+        const post = await this.postRepository.save(
             this.postRepository.create({
                 title,
                 description,
@@ -114,10 +103,21 @@ export class PostService {
                 status,
                 user,
                 category,
-                tags,
-                images
+                tags
             })
         );
+        
+        const postId = post.id;
+        if (fileNames) {
+            for (const fileName of fileNames) {
+                await this.filesService.uploadPostImage(fileName, `${postId}`);
+            }
+        }
+
+        post.content = this.filesService.uploadImageUrlInHtml(post.content, postId);
+        await this.postRepository.save(post);
+
+        return post;
     }
 
     async updateStatus(id: number, status: PostStatus): Promise<Post> {
@@ -144,6 +144,10 @@ export class PostService {
                 }
             ]
         });
+        if (!post) throw new NotFoundException('존재하지 않는 게시물입니다.');
+        const postId: string = post.id.toString();
+        this.filesService.deletePostImages(postId);
+
         await this.postRepository.remove(post);
     }
 }
